+++ KDTree.cpp
@@ -5,13 +5,13 @@
 KDTree::KDTree() : index(-1), valid(false) {}
 
 // בניית עץ מקובץ מטריצה
-KDTree::KDTree(const Eigen::MatrixXd& data) {
+KDTree::KDTree(const Eigen::MatrixXf& data) {
     if (data.rows() == 0) {
         valid = false;
         return;
     }
 
-    std::vector<std::pair<Eigen::Vector3d, int>> pts;
+    std::vector<std::pair<Eigen::Vector3f, int>> pts;
     for (int i = 0; i < data.rows(); ++i) {
         pts.emplace_back(data.row(i).transpose(), i);
     }
@@ -21,7 +21,7 @@ KDTree::KDTree(const Eigen::MatrixXd& data) {
 }
 
 // בניית עץ מרשימת נקודות (נקודת רקורסיה)
-KDTree::KDTree(const std::vector<std::pair<Eigen::Vector3d, int>>& points, int depth) {
+KDTree::KDTree(const std::vector<std::pair<Eigen::Vector3f, int>>& points, int depth) {
     if (points.empty()) {
         valid = false;
         return;
@@ -31,7 +31,7 @@ KDTree::KDTree(const std::vector<std::pair<Eigen::Vector3d, int>>& points, int d
     auto comp = [axis](const auto& a, const auto& b) {
         return a.first[axis] < b.first[axis];
         };
-    std::vector<std::pair<Eigen::Vector3d, int>> sorted_points = points;
+    std::vector<std::pair<Eigen::Vector3f, int>> sorted_points = points;
     std::sort(sorted_points.begin(), sorted_points.end(), comp);
 
     int mid = sorted_points.size() / 2;
@@ -39,8 +39,8 @@ KDTree::KDTree(const std::vector<std::pair<Eigen::Vector3d, int>>& points, int d
     index = sorted_points[mid].second;
     valid = true;
 
-    std::vector<std::pair<Eigen::Vector3d, int>> left_pts(sorted_points.begin(), sorted_points.begin() + mid);
-    std::vector<std::pair<Eigen::Vector3d, int>> right_pts(sorted_points.begin() + mid + 1, sorted_points.end());
+    std::vector<std::pair<Eigen::Vector3f, int>> left_pts(sorted_points.begin(), sorted_points.begin() + mid);
+    std::vector<std::pair<Eigen::Vector3f, int>> right_pts(sorted_points.begin() + mid + 1, sorted_points.end());
 
     if (!left_pts.empty())
         left = std::make_unique<KDTree>(left_pts, depth + 1);
@@ -49,17 +49,17 @@ KDTree::KDTree(const std::vector<std::pair<Eigen::Vector3d, int>>& points, int d
 }
 
 
-void KDTree::nearest(const Eigen::Vector3d& point, int& index, double& dist_sq) const {
+void KDTree::nearest(const Eigen::Vector3f& point, int& index, float& dist_sq) const {
     index = -1;
-    dist_sq = std::numeric_limits<double>::max();
+    dist_sq = std::numeric_limits<float>::max();
     nearestSearch(point, 0, index, dist_sq);
 }
 
-void KDTree::nearestSearch(const Eigen::Vector3d& target, int depth,
-    int& best_idx, double& best_dist_sq) const {
+void KDTree::nearestSearch(const Eigen::Vector3f& target, int depth,
+    int& best_idx, float& best_dist_sq) const {
     if (!valid) return;
 
-    double d = (point - target).squaredNorm();
+    float d = (point - target).squaredNorm();
     if (d < best_dist_sq) {
         best_dist_sq = d;
         best_idx = index;
@@ -80,29 +80,29 @@ void KDTree::nearestSearch(const Eigen::Vector3d& target, int depth,
 
     if (near) near->nearestSearch(target, depth + 1, best_idx, best_dist_sq);
 
-    double diff = target[axis] - point[axis];
+    float diff = target[axis] - point[axis];
     if (diff * diff < best_dist_sq && far) {
         far->nearestSearch(target, depth + 1, best_idx, best_dist_sq);
     }
 }
 
-vector<float> KDTree::radiusSearch(const Eigen::Vector3d& target, float radius) {
+vector<float> KDTree::radiusSearch(const Eigen::Vector3f& target, float radius) {
     std::vector<float> result;
-    double radiusSq = static_cast<double>(radius) * static_cast<double>(radius);
+    float radiusSq = static_cast<float>(radius) * static_cast<float>(radius);
     radiusSearchRec(target, radiusSq, 0, result);
     return result;
 }
 
-void KDTree::radiusSearchRec(const Eigen::Vector3d& target, double radiusSq, int depth, vector<float>& result) {
+void KDTree::radiusSearchRec(const Eigen::Vector3f& target, float radiusSq, int depth, vector<float>& result) {
     if (!valid) return;
 
-    double d = (point - target).squaredNorm();
+    float d = (point - target).squaredNorm();
     if (d <= radiusSq) {
         result.push_back(index);
     }
 
     int axis = depth % 3;
-    double diff = target[axis] - point[axis];
+    float diff = target[axis] - point[axis];
 
     if (diff < 0) {
         if (left) left->radiusSearchRec(target, radiusSq, depth + 1, result);
@@ -114,7 +114,7 @@ void KDTree::radiusSearchRec(const Eigen::Vector3d& target, double radiusSq, int
     }
 }
 
-void KDTree::insert(const Eigen::Vector3d& new_point, int new_index, int depth) {
+void KDTree::insert(const Eigen::Vector3f& new_point, int new_index, int depth) {
     if (!valid) {
         point = new_point;
         index = new_index;
@@ -127,7 +127,7 @@ void KDTree::insert(const Eigen::Vector3d& new_point, int new_index, int depth) 
         if (left)
             left->insert(new_point, new_index, depth + 1);
         else {
-            std::vector<std::pair<Eigen::Vector3d, int>> single = { {new_point, new_index} };
+            std::vector<std::pair<Eigen::Vector3f, int>> single = { {new_point, new_index} };
             left = std::make_unique<KDTree>(single, depth + 1);
         }
     }
@@ -135,7 +135,7 @@ void KDTree::insert(const Eigen::Vector3d& new_point, int new_index, int depth) 
         if (right)
             right->insert(new_point, new_index, depth + 1);
         else {
-            std::vector<std::pair<Eigen::Vector3d, int>> single = { {new_point, new_index} };
+            std::vector<std::pair<Eigen::Vector3f, int>> single = { {new_point, new_index} };
             right = std::make_unique<KDTree>(single, depth + 1);
         }
     }
